requirements.txt
# requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
transformers==4.42.3
torch==2.3.1
pydantic==2.7.4
pydantic-settings==2.3.4
pillow==10.4.0
pytest==8.2.2
